add/refreshtoken

+++ apps/api/src/services/delete.data/pdrd.delete.ts
@@ -1,10 +1,10 @@
-import prisma from "@/prisma";
-import { getPdrdById } from "../existing-data/pdrd.exist";
+import prisma from '@/prisma';
+import { getPdrdById } from '../existing-data/pdrd.exist';
 
 export const deletePdrdById = async (id: string) => {
-    await getPdrdById(id);
-    return await prisma.pickupDeliveryRequest.delete({
-      where: { id },
-    });
-  };
-  
+  const existingPdrd = await getPdrdById(id);
+  if (!existingPdrd) throw new Error('Pickup Delivery Request not found!');
+  return await prisma.pickupDeliveryRequest.delete({
+    where: { id },
+  });
+};
